+++ model/legacy-private/src/main/java/org/keycloak/services/scheduled/OpenIdFederationIdPExpirationTask.java
@@ -4,6 +4,7 @@ import org.jboss.logging.Logger;
 import org.keycloak.models.IdentityProviderModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
+import org.keycloak.protocol.trustchain.TrustChainProcessor;
 import org.keycloak.timer.ScheduledTask;
 import org.keycloak.timer.TimerProvider;
 
@@ -33,9 +34,8 @@ public class OpenIdFederationIdPExpirationTask implements ScheduledTask {
             TimerProvider timer = session.getProvider(TimerProvider.class);
             timer.cancelTaskAndNotify("OpenIdFederationIdPExpirationTask_" + alias);
         } else {
-            //TODO retry for updating IdP
-            idp.setEnabled(false);
-            realm.updateIdentityProvider(idp);
+            TrustChainProcessor trustChainProcessor = session.getProvider(TrustChainProcessor.class, "openid-federation");
+            trustChainProcessor.updateIdP(idp, realm);
         }
     }
 }