+++ server-spi-private/src/main/java/org/keycloak/protocol/trustchain/TrustChainProcessorFactory.java
@@ -1,0 +1,6 @@
+package org.keycloak.protocol.trustchain;
+
+import org.keycloak.provider.ProviderFactory;
+
+public interface TrustChainProcessorFactory extends ProviderFactory<TrustChainProcessor> {
+}