+++ services/src/main/java/org/keycloak/protocol/oidc/endpoints/UserInfoEndpoint.java
@@ -17,6 +17,8 @@
 package org.keycloak.protocol.oidc.endpoints;
 
 import org.jboss.resteasy.annotations.cache.NoCache;
+import org.keycloak.authentication.AuthenticationFlowError;
+import org.keycloak.common.util.UriUtils;
 import org.keycloak.http.HttpRequest;
 import org.keycloak.OAuth2Constants;
 import org.keycloak.TokenCategory;
@@ -43,17 +45,26 @@ import org.keycloak.models.AuthenticatedClientSessionModel;
 import org.keycloak.models.ClientModel;
 import org.keycloak.models.ClientSessionContext;
 import org.keycloak.models.KeycloakSession;
+import org.keycloak.models.OpenIdFederationConfig;
 import org.keycloak.models.RealmModel;
 import org.keycloak.models.UserModel;
 import org.keycloak.models.UserSessionModel;
+import org.keycloak.models.enums.ClientRegistrationTypeEnum;
+import org.keycloak.models.enums.EntityTypeEnum;
 import org.keycloak.protocol.oidc.OIDCAdvancedConfigWrapper;
 import org.keycloak.protocol.oidc.OIDCLoginProtocol;
 import org.keycloak.protocol.oidc.TokenManager;
 import org.keycloak.protocol.oidc.TokenManager.NotBeforeCheck;
 import org.keycloak.representations.AccessToken;
+import org.keycloak.representations.openid_federation.EntityStatement;
+import org.keycloak.representations.openid_federation.RPMetadata;
+import org.keycloak.representations.openid_federation.TrustChainResolution;
+import org.keycloak.services.ErrorResponseException;
 import org.keycloak.services.Urls;
 import org.keycloak.services.clientpolicy.ClientPolicyException;
 import org.keycloak.services.clientpolicy.context.UserInfoRequestContext;
+import org.keycloak.services.clientregistration.ClientRegistrationAuth;
+import org.keycloak.services.clientregistration.openid_federation.OpenIdFederationClientRegistrationService;
 import org.keycloak.services.managers.AppAuthManager;
 import org.keycloak.services.resources.Cors;
 import org.keycloak.services.util.DefaultClientSessionContext;
@@ -72,11 +83,14 @@ import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.HttpHeaders;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.MultivaluedMap;
+import org.keycloak.utils.OpenIdFederationTrustChainProcessor;
+import org.keycloak.utils.OpenIdFederationUtils;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.security.Key;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
  * @author pedroigor
@@ -187,10 +201,21 @@ public class UserInfoEndpoint {
                 throw error.insufficientScope("Missing openid scope");
             }
 
-            clientModel = realm.getClientByClientId(token.getIssuedFor());
-            if (clientModel == null) {
+            String clientId = token.getIssuedFor();
+            clientModel = realm.getClientByClientId(clientId);
+            if (UriUtils.isUri(clientId) && (clientModel == null || !clientModel.isEnabled()) && realm.isOpenIdFederationTypeRegistrationSupported(EntityTypeEnum.OPENID_PROVIDER, ClientRegistrationTypeEnum.AUTOMATIC)) {
+                try {
+                    clientModel = OpenIdFederationUtils.createOrUpdateAutomaticClient(clientId, session, realm);
+                } catch (Exception e) {
+                    event.error(clientModel == null ? Errors.CLIENT_NOT_FOUND : Errors.CLIENT_DISABLED);
+                    throw error.invalidToken(clientModel == null ?"Client not found" : "Client disabled");
+                }
+            } else if (clientModel == null) {
                 event.error(Errors.CLIENT_NOT_FOUND);
                 throw error.invalidToken("Client not found");
+            } else if (!clientModel.isEnabled()) {
+                event.error(Errors.CLIENT_DISABLED);
+                throw error.invalidToken("Client disabled");
             }
 
             cors.allowedOrigins(session, clientModel);
@@ -214,11 +239,6 @@ public class UserInfoEndpoint {
         session.getContext().setClient(clientModel);
 
         event.client(clientModel);
-
-        if (!clientModel.isEnabled()) {
-            event.error(Errors.CLIENT_DISABLED);
-            throw error.invalidToken("Client disabled");
-        }
 
         UserSessionModel userSession = UserSessionUtil.findValidSession(session, realm, token, event, clientModel);