+++ services/src/main/java/org/keycloak/utils/OpenIdFederationTrustChainProcessorFactory.java
@@ -1,0 +1,35 @@
+package org.keycloak.utils;
+
+import org.keycloak.Config;
+import org.keycloak.models.KeycloakSession;
+import org.keycloak.models.KeycloakSessionFactory;
+import org.keycloak.protocol.trustchain.TrustChainProcessor;
+import org.keycloak.protocol.trustchain.TrustChainProcessorFactory;
+
+public class OpenIdFederationTrustChainProcessorFactory implements TrustChainProcessorFactory {
+
+    public static final String PROVIDER_ID = "openid-federation";
+
+    @Override
+    public TrustChainProcessor create(KeycloakSession session) {
+        return new OpenIdFederationTrustChainProcessor(session);
+    }
+
+    @Override
+    public void init(Config.Scope config) {
+    }
+
+    @Override
+    public void postInit(KeycloakSessionFactory factory) {
+    }
+
+    @Override
+    public void close() {
+    }
+
+    @Override
+    public String getId() {
+        return PROVIDER_ID;
+    }
+
+}